safari-600.7-branch/LayoutTests/storage/indexeddb/createObjectStore-name-argument-required-expected.txt

Test IndexedDB createObjectStore name argument is required

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "createObjectStore-name-argument-required.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
db = event.target.result
PASS db.createObjectStore(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
